Modifications to i2c bus to handle call-ahead properly. Fixes to MGC3130 driver. New message definitions.

+++ ManuvrOS/Drivers/i2c-adapter/I2CDevice.cpp
@@ -73,7 +73,7 @@ bool I2CDevice::disassignBusInstance(void) {
 /*
   This function sends the MSB first.
 */
-bool I2CDevice::write16(uint8_t sub_addr, uint16_t dat) {
+bool I2CDevice::write16(int sub_addr, uint16_t dat) {
     if (_bus == NULL) {
       StaticHub::log(__PRETTY_FUNCTION__, 2, "No bus assignment (i2c addr 0x%02x).", _dev_addr);
       return false;
@@ -100,7 +100,7 @@ bool I2CDevice::write8(uint8_t dat) {
 }
 
 
-bool I2CDevice::write8(uint8_t sub_addr, uint8_t dat) {
+bool I2CDevice::write8(int sub_addr, uint8_t dat) {
     if (_bus == NULL) {
       StaticHub::log(__PRETTY_FUNCTION__, 2, "No bus assignment (i2c addr 0x%02x).", _dev_addr);
       return false;
@@ -114,7 +114,7 @@ bool I2CDevice::write8(uint8_t sub_addr, uint8_t dat) {
 
 
 
-bool I2CDevice::writeX(uint8_t sub_addr, uint16_t byte_count, uint8_t *buf) {
+bool I2CDevice::writeX(int sub_addr, uint16_t byte_count, uint8_t *buf) {
     if (_bus == NULL) {
       StaticHub::log(__PRETTY_FUNCTION__, 2, "No bus assignment (i2c addr 0x%02x).", _dev_addr);
       return false;
@@ -129,7 +129,7 @@ bool I2CDevice::writeX(uint8_t sub_addr, uint16_t byte_count, uint8_t *buf) {
 
 
 
-bool I2CDevice::readX(uint8_t sub_addr, uint8_t len, uint8_t *buf) {
+bool I2CDevice::readX(int sub_addr, uint8_t len, uint8_t *buf) {
     if (_bus == NULL) {
       StaticHub::log(__PRETTY_FUNCTION__, 2, "No bus assignment (i2c addr 0x%02x).", _dev_addr);
       return false;
@@ -142,7 +142,7 @@ bool I2CDevice::readX(uint8_t sub_addr, uint8_t len, uint8_t *buf) {
 
 
 
-bool I2CDevice::read8(uint8_t sub_addr) {
+bool I2CDevice::read8(int sub_addr) {
     if (_bus == NULL) {
       StaticHub::log(__PRETTY_FUNCTION__, 2, "No bus assignment (i2c addr 0x%02x).", _dev_addr);
       return false;
